aqua-tests/requirements.txt
delegator.py==0.1.1
pytest==7.1.3
pytest-xdist==2.5.0
pytest-repeat==0.9.1
